Add healthcheck for RabbitMQ node
RabbitMQ healthchecks are hard, as captured by @michaelklishin in
https://github.com/docker-library/rabbitmq/pull/174#issuecomment-452002696
and followed-up in rabbitmq/rabbitmq-cli#292

This is an attempt to simplify the inherent complexity by taking the
first step in what will likely become a new rabbitmqctl diagnostics
command. To follow-up on this idea, join rabbitmq/rabbitmq-cli#292.

Things that we might want to discuss:

* is 30s a reasonable amount to wait before starting a RabbitMQ node for healthines?
* should we account for nodes that can take a long time to boot?
* should we retry longer before considering the node unhealthy?
* is 3s a reasonable command timeout?
* should we improve the active listeners check?

Superceeds docker-library/rabbitmq#174
Related to docker-library/docs#1395