Build RabbitMQ alpha versions instead of release candidates
Alpha releases (a.k.a. edge) are for RabbitMQ users that want to get the
latest RabbitMQ features, as soon as they get committed and pass the
build pipeline, https://ci.rabbitmq.com

The primary reason to consume RabbitMQ alpha releases is because they
are produced continuously, new features and fixes can be tested as they
are addressed. If an issue has been reported and addressed, it will be
first available in an alpha release. Making these releases available via
Docker images allows users to easily test and give feedback.

The secondary reason to switch to RabbitMQ alpha releases is that they
are no less stable than release candidates. All releases, regardless of
type, are produced by the same build pipeline and undergo the same
rigorous testing. The only difference between these two releases types
is timing and scope: alphas are smaller in scope and produced more
frequently.

The last reason for this change is that RabbitMQ release candidate
artefacts get automatically cleaned when a new stable release is
produced. This was first noticed in
https://github.com/docker-library/rabbitmq/pull/297/commits/5a8c3dc714ebe301fadc7572bdb918e9b3bf70b4.
What I've initially thought to be a bug turned out to be "works as
designed".

Lastly, from a base OS perspective, Alpine Edge has the latest
Erlang/OTP minor version. One of our recommendations is
to run the latest Erlang/OTP release for the best TLS support:
http://www.rabbitmq.com/which-erlang.html#supported-version-policy
Alpine Latest ships Erlang/OTP 20.3, while Alpine Edge ships 21.2.

A Debian-based edge version is not available, my expectation is that #297
will result in Ubuntu replacing Debian as the Alpine OS alternative.
To go further, I would challenge the need to have 2 different image
variants for edge versions. What do others think?