Use environment variable to config file mapping
I wanted to submit a more complete pull request, but it is sooo hard to understand everything that is going on in this script that I wasn't confident enough to convert everything to use the new config map. It seems like the approach I'm taking here is a LOT more intuitive and allows any config setting to be mapped from an environment variable. Is this something that you would be interested in doing completely and trying to completely convert to this style? I think it would also end up simplifying this script a lot.