Stop writing env var config to a conf.d config file
In order to be as flexible as possible, I've reverted this back to
leaving the config in rabbitmq.conf. A user may reasonably alter
$RABBITMQ_CONFIG_FILES, at which point any config files in the conf.d
directory won't be picked up.

This now keeps the env var config in $RABBITMQ_CONFIG_FILE, but prints a
deprecation warning with the example config file the user will have to
mount in the future.