Explicitly enable builtin zlib
Until now, the `./configure` option attempted to use shared
(system-provided) zlib. However, the image, at least ubuntu
doesn't have `zlib1g-dev` package installed, so OTP falls back
to using builtin zlib. Therefore, this commit shouldn't change the
contents of the image that gets built, but makes it explicit that it
uses zlib version included in OTP (currently 1.2.12).

The alternative would be to install the missing package and actually
use ubuntu-provided zlib version. However, that would be a change in
behaviour and a negative one - there was a significant performance
improvement included in 1.2.12, that we would have lost.