Fix OpenSSL platform detection, especially for 32bit arches
Thanks to a bug in buildx/buildkit, our CI failed to catch that i386 has been failing to build (for some reason, OpenSSL was detecting it as `linux-x32` instead of `linux-x86`), and our arm32v7 builds were failing for a similar reason (assuming "armv4" instead of v7).

This makes the detection explicit based on the userspace architecture instead.