Set `RUNNING_UNDER_SYSTEMD=true` so `beam.smp` is PID 1
The current Dockerfile / docker-entrypoint.sh results in the following
PIDs within a running RabbitMQ container:

```
PID CMD
  1 /bin/sh /opt/rabbitmq/sbin/rabbitmq-server
 20 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/beam.smp -W w ...
 26 erl_child_setup 1024
 65 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/inet_gethost 4
 66 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/inet_gethost 4
 76 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/epmd -daemon
121 /bin/sh -s rabbit_disk_monitor
```

Note that the `rabbitmq-server` script remains running and is PID 1.

There was a [long discussion](https://github.com/rabbitmq/cluster-operator/discussions/2012)
about what results this particular setup could have in heavily-loaded
k8s environments. This prompted me to look at the `rabbitmq-server`
script and found that the behavior can be controlled via several env
variables:

https://github.com/rabbitmq/rabbitmq-server/blob/main/deps/rabbit/scripts/rabbitmq-server#L97

Most notably, if `RUNNING_UNDER_SYSTEMD` is set to a value, the script
will `exec` the Erlang VM. This PR sets that value, which results in the
following PIDs in a container:

```
PID CMD
  1 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/beam.smp -W w ...
 25 erl_child_setup 1024
 64 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/inet_gethost 4
 65 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/inet_gethost 4
 75 /opt/erlang/lib/erlang/erts-15.2.7.4/bin/epmd -daemon
120 /bin/sh -s rabbit_disk_monitor
```

The Erlang VM already gracefully stops RabbitMQ on `SIGTERM`, so there
is no change in behavior.